web/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
python-multipart>=0.0.9
ruamel.yaml>=0.18.6
watchdog>=4.0.0
